Add 'health' keyword & provide feedback for invalid input in console
r5068
2019-10-03

## DISCUSSION
1. https://github.com/mars-sim/mars-sim/issues/164

## NEW
1. Add "health" keyword for asking a settler how his/her
   health in askPersonRobot() in PersonRobotChatUtils
   and ChatUtil.
   Display the following indices :
   - fatigue
   - thirst
   - hunger
   - energy
   - stress
   - performance
   - surplus ghrelin level
   - surplus leptin level
2. Add hasAnyMatch() in Settlement
   - Checks if any elements from player's input in the
     console would match the name of a vehicle in a
     settlement.

## CHANGE
1. Make BMI figure to have 2 decimals places.
2. Adjust WalkMeta's probability.
3. Split into printDeathNotice() in acquireParty() in
   PartyUtils.
4. Add calling clarify() in various methods.

## FIX
1. Correct returnVehicleList() in Settlement.

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/science/ScientificStudy.java
@@ -527,7 +527,7 @@ public class ScientificStudy implements Serializable, Comparable<ScientificStudy
 	 */
 	public double getCollaborativeResearchWorkTimeCompleted(Person researcher) {
 		if (collaborativeResearchWorkTime.containsKey(researcher.getIdentifier()))
-			return collaborativeResearchWorkTime.get(researcher.getIdentifier());
+			return collaborativeResearchWorkTime.get((Integer)researcher.getIdentifier());
 		else
 			throw new IllegalArgumentException(researcher + " is not a collaborative researcher in this study.");
 	}
@@ -564,7 +564,7 @@ public class ScientificStudy implements Serializable, Comparable<ScientificStudy
 	 */
 	public boolean isCollaborativeResearchCompleted(Person researcher) {
 		if (collaborativeResearchWorkTime.containsKey(researcher.getIdentifier())) {
-			double currentWorkTime = collaborativeResearchWorkTime.get(researcher.getIdentifier());
+			double currentWorkTime = collaborativeResearchWorkTime.get((Integer)researcher.getIdentifier());
 			double requiredWorkTime = getTotalCollaborativeResearchWorkTimeRequired();
 			return (currentWorkTime >= requiredWorkTime);
 		} else
@@ -671,7 +671,7 @@ public class ScientificStudy implements Serializable, Comparable<ScientificStudy
 	 */
 	public double getCollaborativePaperWorkTimeCompleted(Person researcher) {
 		if (collaborativePaperWorkTime.containsKey(researcher.getIdentifier()))
-			return collaborativePaperWorkTime.get(researcher.getIdentifier());
+			return collaborativePaperWorkTime.get((Integer)researcher.getIdentifier());
 		else
 			throw new IllegalArgumentException(researcher + " is not a collaborative researcher in this study.");
 	}
@@ -706,7 +706,7 @@ public class ScientificStudy implements Serializable, Comparable<ScientificStudy
 	 */
 	public boolean isCollaborativePaperCompleted(Person researcher) {
 		if (collaborativePaperWorkTime.containsKey(researcher.getIdentifier())) {
-			double currentWorkTime = collaborativePaperWorkTime.get(researcher.getIdentifier());
+			double currentWorkTime = collaborativePaperWorkTime.get((Integer)researcher.getIdentifier());
 			double requiredWorkTime = getTotalCollaborativePaperWorkTimeRequired();
 			return (currentWorkTime >= requiredWorkTime);
 		} else
@@ -842,7 +842,7 @@ public class ScientificStudy implements Serializable, Comparable<ScientificStudy
 	public MarsClock getLastCollaborativeResearchWorkTime(Person researcher) {
 		MarsClock result = null;
 		if (lastCollaborativeResearchWorkTime.containsKey(researcher.getIdentifier()))
-			result = lastCollaborativeResearchWorkTime.get(researcher.getIdentifier());
+			result = lastCollaborativeResearchWorkTime.get((Integer)researcher.getIdentifier());
 		return result;
 	}
 
@@ -875,7 +875,7 @@ public class ScientificStudy implements Serializable, Comparable<ScientificStudy
 		double result = 0D;
 
 		if (collaborativeAchievementEarned.containsKey(researcher.getIdentifier()))
-			result = collaborativeAchievementEarned.get(researcher.getIdentifier());
+			result = collaborativeAchievementEarned.get((Integer)researcher.getIdentifier());
 
 		return result;
 	}
